+++ index.js
@@ -1,0 +1,16 @@
+const routingDefs = require("./schema/routingDefs");
+const structuralDefs = require("./schema/structuralDefs");
+
+const query = `
+  type Query {
+    _empty: String
+  }
+`;
+
+const mutations = `
+    type Mutation {
+    _empty: String
+  }
+`;
+
+module.exports = [query, mutations, routingDefs, structuralDefs];
