+++ schema/routingDefs.js
@@ -1,159 +1,166 @@
-module.exports = `
-      type RoutingRuleSet {
-          id: ID!
-          routingRules: [RoutingRule]
-          questionPage: QuestionPage
-          else: RoutingDestination
-        }
+module.exports = `  
 
-        type RoutingRule {
-          id: ID!
-          operation: RoutingOperation
-          conditions: [RoutingCondition]
-          goto: RoutingDestination
-        }
+  extend type Query {
+    pagesAffectedByDeletion(pageId: ID!): [Page]!
+    availableRoutingDestinations(pageId: ID!): AvailableRoutingDestinations!
+  }
 
-        enum LogicalDestinations {
-          NextPage
-          EndOfQuestionnaire
-        }
+  type RoutingRuleSet {
+    id: ID!
+    routingRules: [RoutingRule]
+    questionPage: QuestionPage
+    else: RoutingDestination
+  }
 
-        enum AbsoluteDestinationTypes {
-          Section
-          QuestionPage
-        }
+  type RoutingRule {
+    id: ID!
+    operation: RoutingOperation
+    conditions: [RoutingCondition]
+    goto: RoutingDestination
+  }
 
-        union AbsoluteDestinations = QuestionPage | Section
+  enum LogicalDestinations {
+    NextPage
+    EndOfQuestionnaire
+  }
 
-        type AbsoluteDestination {
-          absoluteDestination: AbsoluteDestinations!
-        }
+  enum AbsoluteDestinationTypes {
+    Section
+    QuestionPage
+  }
 
-        type LogicalDestination {
-          id: ID!
-          logicalDestination: LogicalDestinations!
-        }
+  union AbsoluteDestinations = QuestionPage | Section
 
-        union RoutingDestination = AbsoluteDestination | LogicalDestination
+  type AbsoluteDestination {
+    absoluteDestination: AbsoluteDestinations!
+  }
 
-        type AvailableRoutingDestinations {
-          logicalDestinations: [LogicalDestination]!
-          questionPages: [QuestionPage]!
-          sections: [Section]!
-        }
+  type LogicalDestination {
+    id: ID!
+    logicalDestination: LogicalDestinations!
+  }
 
-        type RoutingCondition {
-          id: ID!
-          comparator: RoutingComparator
-          questionPage: QuestionPage
-          answer: Answer
-          routingValue: RoutingConditionValue
-        }
+  union RoutingDestination = AbsoluteDestination | LogicalDestination
 
-        type IDArrayValue {
-          value: [ID]
-        }
+  type AvailableRoutingDestinations {
+    logicalDestinations: [LogicalDestination]!
+    questionPages: [QuestionPage]!
+    sections: [Section]!
+  }
 
-        union RoutingConditionValue = IDArrayValue
+  type RoutingCondition {
+    id: ID!
+    comparator: RoutingComparator
+    questionPage: QuestionPage
+    answer: Answer
+    routingValue: RoutingConditionValue
+  }
 
-        enum RoutingOperation {
-          And
-          Or
-        }
+  type IDArrayValue {
+    value: [ID]
+  }
 
-        enum RoutingComparator {
-          Equal
-        }
-      
-        extend type Mutation {
-          createRoutingRuleSet(input: CreateRoutingRuleSetInput!): RoutingRuleSet
-          updateRoutingRuleSet(input: UpdateRoutingRuleSetInput!): RoutingRuleSet
-          deleteRoutingRuleSet(input: DeleteRoutingRuleSetInput!): RoutingRuleSet
-          resetRoutingRuleSetElse(input: ResetRoutingRuleSetElseInput!): RoutingRuleSet
-        
-          createRoutingRule(input: CreateRoutingRuleInput!): RoutingRule
-          updateRoutingRule(input: UpdateRoutingRuleInput!): RoutingRule
-          deleteRoutingRule(input: DeleteRoutingRuleInput!): RoutingRule
-          undeleteRoutingRule(input: UndeleteRoutingRuleInput!): RoutingRule
-        
-          createRoutingCondition(input: CreateRoutingConditionInput!): RoutingCondition
-          updateRoutingCondition(input: UpdateRoutingConditionInput!): RoutingCondition
-          deleteRoutingCondition(input: DeleteRoutingConditionInput!): RoutingCondition
-        
-          toggleConditionOption(input: ToggleConditionOptionInput!): RoutingConditionValue
-        
-      }
+  union RoutingConditionValue = IDArrayValue
 
-      input CreateRoutingRuleSetInput {
-          questionPageId: ID!
-        }
+  enum RoutingOperation {
+    And
+    Or
+  }
 
-        input UpdateRoutingRuleSetInput {
-          id: ID!
-          else: RoutingDestinationInput!
-        }
+  enum RoutingComparator {
+    Equal
+  }
 
-        input DeleteRoutingRuleSetInput {
-          id: ID!
-        }
+  extend type Mutation {
+    createRoutingRuleSet(input: CreateRoutingRuleSetInput!): RoutingRuleSet
+    updateRoutingRuleSet(input: UpdateRoutingRuleSetInput!): RoutingRuleSet
+    deleteRoutingRuleSet(input: DeleteRoutingRuleSetInput!): RoutingRuleSet
+    resetRoutingRuleSetElse(input: ResetRoutingRuleSetElseInput!): RoutingRuleSet
 
-        input ResetRoutingRuleSetElseInput {
-          id: ID!
-        }
+    createRoutingRule(input: CreateRoutingRuleInput!): RoutingRule
+    updateRoutingRule(input: UpdateRoutingRuleInput!): RoutingRule
+    deleteRoutingRule(input: DeleteRoutingRuleInput!): RoutingRule
+    undeleteRoutingRule(input: UndeleteRoutingRuleInput!): RoutingRule
 
-        input CreateRoutingRuleInput {
-          operation: RoutingOperation!
-          routingRuleSetId: ID!
-        }
+    createRoutingCondition(input: CreateRoutingConditionInput!): RoutingCondition
+    updateRoutingCondition(input: UpdateRoutingConditionInput!): RoutingCondition
+    deleteRoutingCondition(input: DeleteRoutingConditionInput!): RoutingCondition
 
-        input UpdateRoutingRuleInput {
-          id: ID!
-          operation: RoutingOperation
-          goto: RoutingDestinationInput
-        }
+    toggleConditionOption(
+      input: ToggleConditionOptionInput!
+    ): RoutingConditionValue
+  }
 
-        input DeleteRoutingRuleInput {
-          id: ID!
-        }
+  input CreateRoutingRuleSetInput {
+    questionPageId: ID!
+  }
 
-        input UndeleteRoutingRuleInput {
-          id: ID!
-        }
+  input UpdateRoutingRuleSetInput {
+    id: ID!
+    else: RoutingDestinationInput!
+  }
 
-        input CreateRoutingConditionInput {
-          comparator: RoutingComparator!
-          questionPageId: ID!
-          answerId: ID
-          routingRuleId: ID!
-        }
+  input DeleteRoutingRuleSetInput {
+    id: ID!
+  }
 
-        input UpdateRoutingConditionInput {
-          id: ID!
-          questionPageId: ID!
-          answerId: ID
-        }
+  input ResetRoutingRuleSetElseInput {
+    id: ID!
+  }
 
-        input DeleteRoutingConditionInput {
-          id: ID!
-        }
+  input CreateRoutingRuleInput {
+    operation: RoutingOperation!
+    routingRuleSetId: ID!
+  }
 
-        input ToggleConditionOptionInput {
-          optionId: ID
-          conditionId: ID!
-          checked: Boolean!
-        }
+  input UpdateRoutingRuleInput {
+    id: ID!
+    operation: RoutingOperation
+    goto: RoutingDestinationInput
+  }
 
-        input LogicalDestinationInput {
-          destinationType: LogicalDestinations!
-        }
+  input DeleteRoutingRuleInput {
+    id: ID!
+  }
 
-        input AbsoluteDestinationInput {
-          destinationType: AbsoluteDestinationTypes!
-          destinationId: ID!
-        }
+  input UndeleteRoutingRuleInput {
+    id: ID!
+  }
 
-        input RoutingDestinationInput {
-          logicalDestination: LogicalDestinationInput
-          absoluteDestination: AbsoluteDestinationInput
-        }
-    `;
+  input CreateRoutingConditionInput {
+    comparator: RoutingComparator!
+    questionPageId: ID!
+    answerId: ID
+    routingRuleId: ID!
+  }
+
+  input UpdateRoutingConditionInput {
+    id: ID!
+    questionPageId: ID!
+    answerId: ID
+  }
+
+  input DeleteRoutingConditionInput {
+    id: ID!
+  }
+
+  input ToggleConditionOptionInput {
+    optionId: ID
+    conditionId: ID!
+    checked: Boolean!
+  }
+
+  input LogicalDestinationInput {
+    destinationType: LogicalDestinations!
+  }
+
+  input AbsoluteDestinationInput {
+    destinationType: AbsoluteDestinationTypes!
+    destinationId: ID!
+  }
+
+  input RoutingDestinationInput {
+    logicalDestination: LogicalDestinationInput
+    absoluteDestination: AbsoluteDestinationInput
+  }
+`;
